pyproject.toml
[tool.poetry]
name = "lemonhive"
version = "0.1.0"
description = ""
authors = ["Mehran Kader <kmehran.1106@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
pytest = "^7.1.2"
google-cloud-storage = "^2.5.0"
gunicorn = "^20.1.0"
gevent = "^21.12.0"
python-decouple = "^3.6"
flask-marshmallow = "^0.14.0"
pytest-mock = "^3.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.6.0"
mypy = "^0.971"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
